v2.1.0: break change. user/role/dept模块的数据结构修改

+++ mod/user/dao/departmentdao/departmentdao.go
@@ -41,7 +41,7 @@ func New(cascadeType byte, ds ...*sqlkit.DataSource) Dao {
 	return dao
 }
 
-func (dao Dao) ListByParent(id int32) []*model.Department {
+func (dao Dao) ListByParent(id int64) []*model.Department {
 	return dao.Select().Where(squirrel.Eq{"parent": id}).OrderBy("no").OrderBy("id").List()
 }
